Update MainActivity.java
Add ability to connect to hidden SSIDs with --esn hidden option
Add ability  to leave the application on success with --esn leave_app_on_success option

I also added for debug purposes the display of password type and password. I needed this because I have to connect to wifi with special characters that need escaping in order to be well transmitted to the intent ('$' for example).

if the connection fail, the current Supplicant state is displayed in the application.